Clean up actions errors
