+++ transforms/helpers/eo-prop/private/actions/method.ts
@@ -1,10 +1,16 @@
 import { default as j } from 'jscodeshift';
-import type { ClassMethod } from '../../../ast';
+import type { ClassMethod, Collection } from '../../../ast';
+import {
+  findPaths,
+  makeEOActionInfiniteCallAssertion,
+  makeEOActionInfiniteLiteralAssertion,
+} from '../../../ast';
 import { buildActionDecorator } from '../../../decorator-helper';
 import { replaceActionSuperExpressions } from '../../../transform-helper';
 import EOMethodProp from '../method';
+import type { Action } from './index';
 
-export default class ActionMethod extends EOMethodProp {
+export default class ActionMethod extends EOMethodProp implements Action {
   // FIXME: Try to reuse EOMethodProp build?
   override build(): ClassMethod {
     return replaceActionSuperExpressions(
@@ -22,5 +28,29 @@ export default class ActionMethod extends EOMethodProp {
 
   override get isOverridden(): boolean {
     return this.runtimeData.overriddenActions.includes(this.name);
+  }
+
+  get hasInfiniteLoop(): boolean {
+    const { name, value } = this;
+    const collection = j(value.body) as Collection;
+
+    // Occurrences of this.actionName()
+    const isEOActionInfiniteCall = makeEOActionInfiniteCallAssertion(name);
+    const actionCalls = findPaths(
+      collection,
+      j.CallExpression,
+      isEOActionInfiniteCall
+    );
+
+    // Occurrences of this.get('actionName')() or get(this, 'actionName')()
+    const isEOActionInfiniteLiteral =
+      makeEOActionInfiniteLiteralAssertion(name);
+    const actionLiterals = findPaths(
+      collection,
+      j.StringLiteral,
+      isEOActionInfiniteLiteral
+    );
+
+    return actionLiterals.length > 0 || actionCalls.length > 0;
   }
 }
